ADD: fix howitworks section;
